Remove daemon support
This is has a poor error handling.  Qt event loop can't be try/catched by
Python code, so a sys.exepthook is needed, making it hard to handle errors and
clean up the socket file.  The whole socket reading/writing actually seems to
impact startup time.

Nowadays it seems to be ok to open the program every time we need it, no need to
load Qt stuff in a separate, long-running daemon.