Implement rudimentary undo/redo in JS layer
This is poorly implemented though, since this is implemented in JS while some
useful state is in the Python stuff (like the text of the selected entry).
Because of this, not everything that could be undoable is actually undoable.

Maybe everything should be moved to Python.

Anyway, this is a start.